requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pillow==11.3.0
pydantic==2.11.7
httpx==0.28.1
pytest==8.4.1
websockets==15.0.1
